Housekeeping - Move all date helpers into single
We wanted to bring in `today()` from water-abstraction-system so were looking at creating a new helper. We figured it might be useful to bring in a few other date helpers such as `compareDates()`, `yesterday()` etc, but putting them in their own modules seemed overkill.

So, we put them in a single module, and then realised our existing helpers are all 'date helpers'. Moving them into the same module made sense and also meant we could deal with the fact our existing helpers were not following our file naming conventions.
