+++ cypress/support/scenarios/one-return-requirement-quarterly-returns.js
@@ -1,4 +1,3 @@
-import currentFinancialYear from '../helpers/currentFinancialYear.js'
 import licence from '../fixture-builder/licence.js'
 import licenceVersionPurposes from '../fixture-builder/licence-version-purposes.js'
 import points from '../fixture-builder/points.js'
@@ -6,6 +5,7 @@ import returnLogs from '../fixture-builder/return-logs.js'
 import returnRequirements from '../fixture-builder/return-requirements.js'
 import returnRequirementPoints from '../fixture-builder/return-requirement-points.js'
 import returnVersion from '../fixture-builder/return-version.js'
+import { currentFinancialYear } from '../helpers/date.helpers.js'
 
 export default function () {
   const dataModel = {
